feature(ui): absence converted to atomic design
