+++ apps/web/src/components/migration.md
@@ -1,0 +1,3 @@
+## Absence
+70/100
+Charts, Edit
